+++ MobileMvxApp/BLE.Client/PagesViewModelsSet/ViewModelMainMenu.cs
@@ -209,11 +209,11 @@ namespace BLE.Client.ViewModels
                     }
                 }
 
-                uint portNum = BleMvxApplication._reader.rfid.GetAntennaPort();
+                int portNum = BleMvxApplication._reader.rfid.GetAntennaPort();
                 for (uint cnt = 0; cnt < portNum; cnt++)
                 {
                     BleMvxApplication._reader.rfid.SetAntennaPortState(cnt, BleMvxApplication._config.RFID_AntennaEnable[cnt] ? CSLibrary.Constants.AntennaPortState.ENABLED : CSLibrary.Constants.AntennaPortState.DISABLED);
-                    BleMvxApplication._reader.rfid.SetPowerLevel(BleMvxApplication._config.RFID_Antenna_Power[cnt], cnt);
+                    BleMvxApplication._reader.rfid.SetPowerLevel(BleMvxApplication._config.RFID_Antenna_Power[cnt], (int)cnt);
                     BleMvxApplication._reader.rfid.SetInventoryDuration(BleMvxApplication._config.RFID_Antenna_Dwell[cnt], cnt);
                 }
